x-likes/requirements.txt
pandas
pillow
python-dotenv
selenium
webdriver-manager
